cpu: Remove global job queue
Instead, target a specific CPU for a global job at queuing time.

This will allow us to wake up the target using an interrupt when
implementing nap mode.

The algorithm used is to look for idle primary threads first, then
idle secondaries, and finally the less loaded thread. If nothing can
be found, we fallback to a synchronous call.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>