occ: Disable OCC on UV systems
On Ultravisor enabled system, temporarily disable OCC irq paths for
shared memory updates when OCC is up, so as to boot the system.

We also disable the OPAL-OCC command response interface which is
required for Sensors. This will be enabled at a later point when we
can read the OCC Common area residing in secure memory region.

Disable stop levels if they are enabled in the SPIRA.

Skip pstates_init and occ_sensor_init as well.

Signed-off-by: Ram Pai <linuxram@us.ibm.com>
[merged all OCC fixes into a single fix]
Signed-off-by: Satheesh Rajendran <sathnaga@linux.vnet.ibm.com>
[Reduce occ debug log flood]
Signed-off-by: Shilpasri G Bhat <shilpa.bhat@linux.vnet.ibm.com>
Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
Signed-off-by: Ryan Grimm <grimm@linux.ibm.com>